🔧 Fix import path issues from component reorganization
- Fixed all import paths after component reorganization
- Added TypeScript window type definitions for __TAURI__ property
- Removed unused parameter in AppFooter
- All components now correctly import from their new locations
- Build and type checking now passes successfully
