+++ src/types/window.d.ts
@@ -1,0 +1,8 @@
+// Type definitions for Tauri window extensions
+declare global {
+  interface Window {
+    __TAURI__?: boolean;
+  }
+}
+
+export {};
